CPU: Rewrote instruction decoding
Take decision to read whole byte stream in method
cpu_decodeOperands.
Nico explained that each instruction itself
determines which operand works as source and
destination. In the MOD_RM byte the first operand
(bit 5..3) is always a register. The second
one (bit 2..0) maybe a register or memory address.
As a consequence it makes sense to determine
instantly if a SIB byte must be read and/or
a displacement and or an immidiate value. The
return value of cpu_decodeOperands is a struct
that contains the final address of operand 1 and 2.